skip_invalid_loci option for import methods
- Fixes #3691
