+++ python/hail/tests/test_methods.py
@@ -1170,16 +1170,6 @@ class Tests(unittest.TestCase):
 
         self.assertTrue(men2.filter(men2.s == 'Dtr1')._same(men.filter(men.s == 'Dtr1')))
 
-    def test_export_vcf(self):
-        dataset = hl.import_vcf(resource('sample.vcf.bgz'))
-        vcf_metadata = hl.get_vcf_metadata(resource('sample.vcf.bgz'))
-        hl.export_vcf(dataset, '/tmp/sample.vcf', metadata=vcf_metadata)
-        dataset_imported = hl.import_vcf('/tmp/sample.vcf')
-        self.assertTrue(dataset._same(dataset_imported))
-
-        metadata_imported = hl.get_vcf_metadata('/tmp/sample.vcf')
-        self.assertDictEqual(vcf_metadata, metadata_imported)
-
     def test_concordance(self):
         dataset = self.get_dataset()
         glob_conc, cols_conc, rows_conc = hl.concordance(dataset, dataset)
@@ -1204,93 +1194,6 @@ class Tests(unittest.TestCase):
 
         cols_conc.write('/tmp/foo.kt', overwrite=True)
         rows_conc.write('/tmp/foo.kt', overwrite=True)
-
-    def test_import_table_force_bgz(self):
-        f = utils.new_temp_file(suffix=".bgz")
-        t = utils.range_table(10, 5)
-        t.export(f)
-
-        f2 = utils.new_temp_file(suffix=".gz")
-        utils.run_command(["cp", utils.uri_path(f), utils.uri_path(f2)])
-        t2 = hl.import_table(f2, force_bgz=True, impute=True).key_by('idx')
-        self.assertTrue(t._same(t2))
-
-    def test_import_locus_intervals(self):
-        interval_file = resource('annotinterall.interval_list')
-        t = hl.import_locus_intervals(interval_file, reference_genome='GRCh37')
-        nint = t.count()
-
-        i = 0
-        with open(interval_file) as f:
-            for line in f:
-                if len(line.strip()) != 0:
-                    i += 1
-        self.assertEqual(nint, i)
-        self.assertEqual(t.interval.dtype.point_type, hl.tlocus('GRCh37'))
-
-        tmp_file = utils.new_temp_file(prefix="test", suffix="interval_list")
-        start = t.interval.start
-        end = t.interval.end
-        (t
-         .key_by(interval=hl.locus_interval(start.contig, start.position, end.position, True, True))
-         .select()
-         .export(tmp_file, header=False))
-
-        t2 = hl.import_locus_intervals(tmp_file)
-
-        self.assertTrue(t.select()._same(t2))
-
-    def test_import_locus_intervals_no_reference_specified(self):
-        interval_file = resource('annotinterall.interval_list')
-        t = hl.import_locus_intervals(interval_file, reference_genome=None)
-        self.assertTrue(t.count() == 2)
-        self.assertEqual(t.interval.dtype.point_type, hl.tstruct(contig=hl.tstr, position=hl.tint32))
-
-    def test_import_locus_intervals_badly_defined_intervals(self):
-        interval_file = resource('example3.interval_list')
-        t = hl.import_locus_intervals(interval_file, reference_genome='GRCh37', skip_invalid_intervals=True)
-        self.assertTrue(t.count() == 21)
-
-        t = hl.import_locus_intervals(interval_file, reference_genome=None, skip_invalid_intervals=True)
-        self.assertTrue(t.count() == 22)
-
-    def test_import_bed(self):
-        bed_file = resource('example1.bed')
-        bed = hl.import_bed(bed_file, reference_genome='GRCh37')
-
-        nbed = bed.count()
-        i = 0
-        with open(bed_file) as f:
-            for line in f:
-                if len(line.strip()) != 0:
-                    try:
-                        int(line.split()[0])
-                        i += 1
-                    except:
-                        pass
-        self.assertEqual(nbed, i)
-
-        self.assertEqual(bed.interval.dtype.point_type, hl.tlocus('GRCh37'))
-
-        bed_file = resource('example2.bed')
-        t = hl.import_bed(bed_file, reference_genome='GRCh37')
-        self.assertEqual(t.interval.dtype.point_type, hl.tlocus('GRCh37'))
-        self.assertTrue(list(t.key.dtype) == ['interval'])
-        self.assertTrue(list(t.row.dtype) == ['interval','target'])
-
-    def test_import_bed_no_reference_specified(self):
-        bed_file = resource('example1.bed')
-        t = hl.import_bed(bed_file, reference_genome=None)
-        self.assertTrue(t.count() == 3)
-        self.assertEqual(t.interval.dtype.point_type, hl.tstruct(contig=hl.tstr, position=hl.tint32))
-
-    def test_import_bed_badly_defined_intervals(self):
-        bed_file = resource('example4.bed')
-        t = hl.import_bed(bed_file, reference_genome='GRCh37', skip_invalid_intervals=True)
-        self.assertTrue(t.count() == 3)
-
-        t = hl.import_bed(bed_file, reference_genome=None, skip_invalid_intervals=True)
-        self.assertTrue(t.count() == 4)
 
     def test_annotate_intervals(self):
         ds = self.get_dataset()
@@ -1324,157 +1227,6 @@ class Tests(unittest.TestCase):
 
         self.assertTrue(ds.annotate_rows(target = interval_list2[ds.locus].target).rows()
                         ._same(ds.annotate_rows(target = bed2[ds.locus].target).rows()))
-
-    def test_import_fam(self):
-        fam_file = resource('sample.fam')
-        nfam = hl.import_fam(fam_file).count()
-        i = 0
-        with open(fam_file) as f:
-            for line in f:
-                if len(line.strip()) != 0:
-                    i += 1
-        self.assertEqual(nfam, i)
-
-    def test_export_plink(self):
-        vcf_file = resource('sample.vcf')
-        mt = hl.split_multi_hts(hl.import_vcf(vcf_file, min_partitions=10))
-
-        split_vcf_file = utils.uri_path(utils.new_temp_file())
-        hl_output = utils.uri_path(utils.new_temp_file())
-        plink_output = utils.uri_path(utils.new_temp_file())
-        merge_output = utils.uri_path(utils.new_temp_file())
-
-        hl.export_vcf(mt, split_vcf_file)
-        hl.export_plink(mt, hl_output)
-
-        utils.run_command(["plink", "--vcf", split_vcf_file,
-                           "--make-bed", "--out", plink_output,
-                           "--const-fid", "--keep-allele-order"])
-
-        data = []
-        with open(utils.uri_path(plink_output + ".bim")) as file:
-            for line in file:
-                row = line.strip().split()
-                row[1] = ":".join([row[0], row[3], row[5], row[4]])
-                data.append("\t".join(row) + "\n")
-
-        with open(plink_output + ".bim", 'w') as f:
-            f.writelines(data)
-
-        utils.run_command(["plink", "--bfile", plink_output,
-                           "--bmerge", hl_output, "--merge-mode",
-                           "6", "--out", merge_output])
-
-        same = True
-        with open(merge_output + ".diff") as f:
-            for line in f:
-                row = line.strip().split()
-                if row != ["SNP", "FID", "IID", "NEW", "OLD"]:
-                    same = False
-                    break
-
-        self.assertTrue(same)
-
-    def test_export_plink_exprs(self):
-        ds = self.get_dataset()
-        fam_mapping = {'f0': 'fam_id', 'f1': 'ind_id', 'f2': 'pat_id', 'f3': 'mat_id',
-                       'f4': 'is_female', 'f5': 'pheno'}
-        bim_mapping = {'f0': 'contig', 'f1': 'varid', 'f2': 'cm_position',
-                       'f3': 'position', 'f4': 'a1', 'f5': 'a2'}
-
-        # Test default arguments
-        out1 = utils.new_temp_file()
-        hl.export_plink(ds, out1)
-        fam1 = (hl.import_table(out1 + '.fam', no_header=True, impute=False, missing="")
-                .rename(fam_mapping))
-        bim1 = (hl.import_table(out1 + '.bim', no_header=True, impute=False)
-                .rename(bim_mapping))
-
-        self.assertTrue(fam1.all((fam1.fam_id == "0") & (fam1.pat_id == "0") &
-                                 (fam1.mat_id == "0") & (fam1.is_female == "0") &
-                                 (fam1.pheno == "NA")))
-        self.assertTrue(bim1.all((bim1.varid == bim1.contig + ":" + bim1.position + ":" + bim1.a2 + ":" + bim1.a1) &
-                                 (bim1.cm_position == "0.0")))
-
-        # Test non-default FAM arguments
-        out2 = utils.new_temp_file()
-        hl.export_plink(ds, out2, ind_id=ds.s, fam_id=ds.s, pat_id="nope",
-                        mat_id="nada", is_female=True, pheno=False)
-        fam2 = (hl.import_table(out2 + '.fam', no_header=True, impute=False, missing="")
-                .rename(fam_mapping))
-
-        self.assertTrue(fam2.all((fam2.fam_id == fam2.ind_id) & (fam2.pat_id == "nope") &
-                                 (fam2.mat_id == "nada") & (fam2.is_female == "2") &
-                                 (fam2.pheno == "1")))
-
-        # Test quantitative phenotype
-        out3 = utils.new_temp_file()
-        hl.export_plink(ds, out3, ind_id=ds.s, pheno=hl.float64(hl.len(ds.s)))
-        fam3 = (hl.import_table(out3 + '.fam', no_header=True, impute=False, missing="")
-                .rename(fam_mapping))
-
-        self.assertTrue(fam3.all((fam3.fam_id == "0") & (fam3.pat_id == "0") &
-                                 (fam3.mat_id == "0") & (fam3.is_female == "0") &
-                                 (fam3.pheno != "0") & (fam3.pheno != "NA")))
-
-        # Test non-default BIM arguments
-        out4 = utils.new_temp_file()
-        hl.export_plink(ds, out4, varid="hello", cm_position=100)
-        bim4 = (hl.import_table(out4 + '.bim', no_header=True, impute=False)
-                .rename(bim_mapping))
-
-        self.assertTrue(bim4.all((bim4.varid == "hello") & (bim4.cm_position == "100.0")))
-
-        # Test call expr
-        out5 = utils.new_temp_file()
-        ds_call = ds.annotate_entries(gt_fake=hl.call(0, 0))
-        hl.export_plink(ds_call, out5, call=ds_call.gt_fake)
-        ds_all_hom_ref = hl.import_plink(out5 + '.bed', out5 + '.bim', out5 + '.fam')
-        nerrors = ds_all_hom_ref.aggregate_entries(agg.count_where(~ds_all_hom_ref.GT.is_hom_ref()))
-        self.assertTrue(nerrors == 0)
-
-        # Test white-space in FAM id expr raises error
-        with self.assertRaisesRegex(TypeError, "has spaces in the following values:"):
-            hl.export_plink(ds, utils.new_temp_file(), mat_id="hello world")
-
-        # Test white-space in varid expr raises error
-        with self.assertRaisesRegex(utils.FatalError, "no white space allowed:"):
-            hl.export_plink(ds, utils.new_temp_file(), varid="hello world")
-
-    def test_export_gen(self):
-        gen = hl.import_gen(resource('example.gen'),
-                            sample_file=resource('example.sample'),
-                            contig_recoding={"01": "1"},
-                            reference_genome='GRCh37',
-                            min_partitions=3)
-
-        file = '/tmp/test_export_gen'
-        hl.export_gen(gen, file)
-        gen2 = hl.import_gen(file + '.gen',
-                             sample_file=file + '.sample',
-                             reference_genome='GRCh37',
-                             min_partitions=3)
-
-        self.assertTrue(gen._same(gen2, tolerance=3E-4, absolute=True))
-
-    def test_export_gen_exprs(self):
-        gen = hl.import_gen(resource('example.gen'),
-                            sample_file=resource('example.sample'),
-                            contig_recoding={"01": "1"},
-                            reference_genome='GRCh37',
-                            min_partitions=3).add_col_index().add_row_index()
-
-        out1 = utils.new_temp_file()
-        hl.export_gen(gen, out1, id1=hl.str(gen.col_idx), id2=hl.str(gen.col_idx), missing=0.5,
-                      varid=hl.str(gen.row_idx), rsid=hl.str(gen.row_idx), gp=[0.0, 1.0, 0.0])
-
-        in1 = (hl.import_gen(out1 + '.gen', sample_file=out1 + '.sample', min_partitions=3)
-               .add_col_index()
-               .add_row_index())
-        self.assertTrue(in1.aggregate_entries(agg.fraction(in1.GP == [0.0, 1.0, 0.0])) == 1.0)
-        self.assertTrue(in1.aggregate_rows(agg.fraction((in1.varid == hl.str(in1.row_idx)) &
-                                                        (in1.rsid == hl.str(in1.row_idx)))) == 1.0)
-        self.assertTrue(in1.aggregate_cols(agg.fraction((in1.s == hl.str(in1.col_idx)))))
 
     def test_tdt(self):
         pedigree = hl.Pedigree.read(resource('tdt.fam'))
@@ -1761,252 +1513,6 @@ class Tests(unittest.TestCase):
                 x=hl.pl_dosage(ds.PL),
                 covariates=[ds.cov.Cov1, ds.cov.Cov2],
                 logistic=True).count()
-
-    def test_import_gen(self):
-        gen = hl.import_gen(resource('example.gen'),
-                            sample_file=resource('example.sample'),
-                            contig_recoding={"01": "1"},
-                            reference_genome = 'GRCh37').rows()
-        self.assertTrue(gen.all(gen.locus.contig == "1"))
-        self.assertEqual(gen.count(), 199)
-        self.assertEqual(gen.locus.dtype, hl.tlocus('GRCh37'))
-
-    def test_import_gen_no_reference_specified(self):
-        gen = hl.import_gen(resource('example.gen'),
-                            sample_file=resource('example.sample'),
-                            reference_genome=None)
-
-        self.assertTrue(gen.locus.dtype == hl.tstruct(contig=hl.tstr, position=hl.tint32))
-        self.assertEqual(gen.count_rows(), 199)
-
-    def test_import_bgen(self):
-        hl.index_bgen(resource('example.v11.bgen'))
-
-        bgen_rows = hl.import_bgen(resource('example.v11.bgen'),
-                                   entry_fields=['GT', 'GP'],
-                                   sample_file=resource('example.sample'),
-                                   contig_recoding={'01': '1'},
-                                   reference_genome='GRCh37').rows()
-        self.assertTrue(bgen_rows.all(bgen_rows.locus.contig == '1'))
-        self.assertEqual(bgen_rows.count(), 199)
-
-        hl.index_bgen(resource('example.8bits.bgen'))
-
-        bgen = hl.import_bgen(resource('example.8bits.bgen'),
-                              entry_fields=['dosage'],
-                              contig_recoding={'01': '1'},
-                              reference_genome='GRCh37')
-        self.assertEqual(bgen.entry.dtype, hl.tstruct(dosage=hl.tfloat64))
-
-        bgen = hl.import_bgen(resource('example.8bits.bgen'),
-                              entry_fields=['GT', 'GP'],
-                              sample_file=resource('example.sample'),
-                              contig_recoding={'01': '1'},
-                              reference_genome='GRCh37')
-        self.assertEqual(bgen.entry.dtype, hl.tstruct(GT=hl.tcall, GP=hl.tarray(hl.tfloat64)))
-        self.assertEqual(bgen.count_rows(), 199)
-
-        hl.index_bgen(resource('example.10bits.bgen'))
-        bgen = hl.import_bgen(resource('example.10bits.bgen'),
-                              entry_fields=['GT', 'GP', 'dosage'],
-                              contig_recoding={'01': '1'},
-                              reference_genome='GRCh37')
-        self.assertEqual(bgen.entry.dtype, hl.tstruct(GT=hl.tcall, GP=hl.tarray(hl.tfloat64), dosage=hl.tfloat64))
-        self.assertEqual(bgen.locus.dtype, hl.tlocus('GRCh37'))
-
-    def test_import_bgen_no_entry_fields(self):
-        hl.index_bgen(resource('example.v11.bgen'))
-
-        bgen = hl.import_bgen(resource('example.v11.bgen'),
-                              entry_fields=[],
-                              sample_file=resource('example.sample'),
-                              contig_recoding={'01': '1'},
-                              reference_genome='GRCh37')
-        bgen._jvds.typecheck()
-
-    def test_import_bgen_no_reference_specified(self):
-        bgen = hl.import_bgen(resource('example.10bits.bgen'),
-                              entry_fields=['GT', 'GP', 'dosage'],
-                              contig_recoding={'01': '1'},
-                              reference_genome=None)
-        self.assertTrue(bgen.locus.dtype == hl.tstruct(contig=hl.tstr, position=hl.tint32))
-        self.assertEqual(bgen.count_rows(), 199)
-
-    def test_import_vcf(self):
-        vcf = hl.split_multi_hts(
-            hl.import_vcf(resource('sample2.vcf'),
-                          reference_genome=hl.get_reference('GRCh38'),
-                          contig_recoding={"22": "chr22"}))
-
-        vcf_table = vcf.rows()
-        self.assertTrue(vcf_table.all(vcf_table.locus.contig == "chr22"))
-        self.assertTrue(vcf.locus.dtype, hl.tlocus('GRCh37'))
-
-    def test_import_vcf_no_reference_specified(self):
-        vcf = hl.import_vcf(resource('sample2.vcf'),
-                            reference_genome=None)
-        self.assertTrue(vcf.locus.dtype == hl.tstruct(contig=hl.tstr, position=hl.tint32))
-        self.assertEqual(vcf.count_rows(), 735)
-
-    def test_import_vcf_bad_reference_allele(self):
-        vcf = hl.import_vcf(resource('invalid_base.vcf'))
-        self.assertEqual(vcf.count_rows(), 1)
-
-    def test_import_vcf_flags_are_defined(self):
-        # issue 3277
-        t = hl.import_vcf(resource('sample.vcf')).rows()
-        self.assertTrue(t.all(hl.is_defined(t.info.NEGATIVE_TRAIN_SITE) &
-                              hl.is_defined(t.info.POSITIVE_TRAIN_SITE) &
-                              hl.is_defined(t.info.DB) &
-                              hl.is_defined(t.info.DS)))
-
-    def test_import_vcf_can_import_float_array_format(self):
-        mt = hl.import_vcf(resource('floating_point_array.vcf'))
-        self.assertTrue(mt.aggregate_entries(hl.agg.all(mt.numeric_array == [1.5, 2.5])))
-
-    def test_import_vcf_can_import_negative_numbers(self):
-        mt = hl.import_vcf(resource('negative_format_fields.vcf'))
-        self.assertTrue(mt.aggregate_entries(hl.agg.all(mt.negative_int == -1) &
-                                             hl.agg.all(mt.negative_float == -1.5) &
-                                             hl.agg.all(mt.negative_int_array == [-1, -2]) &
-                                             hl.agg.all(mt.negative_float_array == [-0.5, -1.5])))
-
-    def test_import_vcf_missing_info_field_elements(self):
-        mt = hl.import_vcf(resource('missingInfoArray.vcf'), reference_genome='GRCh37', array_elements_required=False)
-        mt = mt.select_rows(FOO=mt.info.FOO, BAR=mt.info.BAR)
-        expected = hl.Table.parallelize([{'locus': hl.Locus('X', 16050036), 'alleles': ['A', 'C'],
-                                          'FOO': [1, None], 'BAR': [2, None, None]},
-                                         {'locus': hl.Locus('X', 16061250), 'alleles': ['T', 'A', 'C'],
-                                          'FOO': [None, 2, None], 'BAR': [None, 1.0, None]}],
-                                        hl.tstruct(locus=hl.tlocus('GRCh37'), alleles=hl.tarray(hl.tstr),
-                                                   FOO=hl.tarray(hl.tint), BAR=hl.tarray(hl.tfloat64)),
-                                        key=['locus', 'alleles'])
-        self.assertTrue(mt.rows()._same(expected))
-
-    def test_import_vcf_missing_format_field_elements(self):
-        mt = hl.import_vcf(resource('missingFormatArray.vcf'), reference_genome='GRCh37', array_elements_required=False)
-        mt = mt.select_rows().select_entries('AD', 'PL')
-
-        expected = hl.Table.parallelize([{'locus': hl.Locus('X', 16050036), 'alleles': ['A', 'C'], 's': 'C1046::HG02024',
-                                          'AD': [None, None], 'PL': [0, None, 180]},
-                                         {'locus': hl.Locus('X', 16050036), 'alleles': ['A', 'C'], 's': 'C1046::HG02025',
-                                          'AD': [None, 6], 'PL': [70, None]},
-                                         {'locus': hl.Locus('X', 16061250), 'alleles': ['T', 'A', 'C'], 's': 'C1046::HG02024',
-                                          'AD': [0, 0, None], 'PL': [396, None, None, 33, None, 0]},
-                                         {'locus': hl.Locus('X', 16061250), 'alleles': ['T', 'A', 'C'], 's': 'C1046::HG02025',
-                                          'AD': [0, 0, 9], 'PL': [None, None, None]}],
-                                        hl.tstruct(locus=hl.tlocus('GRCh37'), alleles=hl.tarray(hl.tstr), s=hl.tstr,
-                                                   AD=hl.tarray(hl.tint), PL=hl.tarray(hl.tint)),
-                                        key=['locus', 'alleles', 's'])
-
-        self.assertTrue(mt.entries()._same(expected))
-
-    def test_export_import_plink_same(self):
-        mt = self.get_dataset()
-        mt = mt.select_rows(rsid=hl.delimit([mt.locus.contig, hl.str(mt.locus.position), mt.alleles[0], mt.alleles[1]], ':'),
-                            cm_position=15.0)
-        mt = mt.select_cols(fam_id=hl.null(hl.tstr), pat_id=hl.null(hl.tstr), mat_id=hl.null(hl.tstr),
-                            is_female=hl.null(hl.tbool), is_case=hl.null(hl.tbool))
-        mt = mt.select_entries('GT')
-
-        bfile = '/tmp/test_import_export_plink'
-        hl.export_plink(mt, bfile, ind_id=mt.s, cm_position=mt.cm_position)
-
-        mt_imported = hl.import_plink(bfile + '.bed', bfile + '.bim', bfile + '.fam',
-                                      a2_reference=True, reference_genome='GRCh37')
-        self.assertTrue(mt._same(mt_imported))
-        self.assertTrue(mt.aggregate_rows(hl.agg.all(mt.cm_position == 15.0)))
-
-    def test_import_plink_empty_fam(self):
-        mt = self.get_dataset().drop_cols()
-        bfile = '/tmp/test_empty_fam'
-        hl.export_plink(mt, bfile, ind_id=mt.s)
-        with self.assertRaisesRegex(utils.FatalError, "Empty .fam file"):
-            hl.import_plink(bfile + '.bed', bfile + '.bim', bfile + '.fam')
-
-    def test_import_plink_empty_bim(self):
-        mt = self.get_dataset().drop_rows()
-        bfile = '/tmp/test_empty_bim'
-        hl.export_plink(mt, bfile, ind_id=mt.s)
-        with self.assertRaisesRegex(utils.FatalError, ".bim file does not contain any variants"):
-            hl.import_plink(bfile + '.bed', bfile + '.bim', bfile + '.fam')
-
-    def test_import_plink_a1_major(self):
-        mt = self.get_dataset()
-        bfile = '/tmp/sample_plink'
-        hl.export_plink(mt, bfile, ind_id=mt.s)
-
-        def get_data(a2_reference):
-            mt_imported = hl.import_plink(bfile + '.bed', bfile + '.bim',
-                                          bfile + '.fam', a2_reference=a2_reference)
-            return (hl.variant_qc(mt_imported)
-                    .rows()
-                    .key_by('rsid'))
-
-        a2 = get_data(a2_reference=True)
-        a1 = get_data(a2_reference=False)
-
-        j = (a2.annotate(a1_alleles=a1[a2.rsid].alleles, a1_vqc=a1[a2.rsid].variant_qc)
-             .rename({'variant_qc': 'a2_vqc', 'alleles': 'a2_alleles'}))
-
-        self.assertTrue(j.all((j.a1_alleles[0] == j.a2_alleles[1]) &
-                              (j.a1_alleles[1] == j.a2_alleles[0]) &
-                              (j.a1_vqc.n_not_called == j.a2_vqc.n_not_called) &
-                              (j.a1_vqc.n_het == j.a2_vqc.n_het) &
-                              (j.a1_vqc.homozygote_count[0] == j.a2_vqc.homozygote_count[1]) &
-                              (j.a1_vqc.homozygote_count[1] == j.a2_vqc.homozygote_count[0])))
-
-    def test_import_plink_contig_recoding_w_reference(self):
-        vcf = hl.split_multi_hts(
-            hl.import_vcf(resource('sample2.vcf'),
-                          reference_genome=hl.get_reference('GRCh38'),
-                          contig_recoding={"22": "chr22"}))
-
-        hl.export_plink(vcf, '/tmp/sample_plink')
-
-        bfile = '/tmp/sample_plink'
-        plink = hl.import_plink(
-            bfile + '.bed', bfile + '.bim', bfile + '.fam',
-            a2_reference=True,
-            contig_recoding={'chr22': '22'},
-            reference_genome='GRCh37').rows()
-        self.assertTrue(plink.all(plink.locus.contig == "22"))
-        self.assertEqual(vcf.count_rows(), plink.count())
-        self.assertTrue(plink.locus.dtype, hl.tlocus('GRCh37'))
-
-    def test_import_plink_no_reference_specified(self):
-        bfile = resource('fastlmmTest')
-        plink = hl.import_plink(bfile + '.bed', bfile + '.bim', bfile + '.fam',
-                                reference_genome=None)
-        self.assertTrue(plink.locus.dtype == hl.tstruct(contig=hl.tstr, position=hl.tint32))
-
-    def test_import_matrix_table(self):
-        mt = hl.import_matrix_table(doctest_resource('matrix1.tsv'),
-                                    row_fields={'Barcode': hl.tstr, 'Tissue': hl.tstr, 'Days': hl.tfloat32})
-        self.assertEqual(mt['Barcode']._indices, mt._row_indices)
-        self.assertEqual(mt['Tissue']._indices, mt._row_indices)
-        self.assertEqual(mt['Days']._indices, mt._row_indices)
-        self.assertEqual(mt['col_id']._indices, mt._col_indices)
-        self.assertEqual(mt['row_id']._indices, mt._row_indices)
-
-        mt.count()
-
-        row_fields = {'f0': hl.tstr, 'f1': hl.tstr, 'f2': hl.tfloat32}
-        hl.import_matrix_table(doctest_resource('matrix2.tsv'),
-                               row_fields=row_fields, row_key=[]).count()
-        hl.import_matrix_table(doctest_resource('matrix3.tsv'),
-                               row_fields=row_fields,
-                               no_header=True).count()
-        hl.import_matrix_table(doctest_resource('matrix3.tsv'),
-                               row_fields=row_fields,
-                               no_header=True,
-                               row_key=[]).count()
-        self.assertRaises(hl.utils.FatalError,
-                     hl.import_matrix_table,
-                          doctest_resource('matrix3.tsv'),
-                     row_fields=row_fields,
-                     no_header=True,
-                     row_key=['foo'])
 
     def test_de_novo(self):
         mt = hl.import_vcf(resource('denovo.vcf'))
