+++ src/main/scala/is/hail/HailContext.scala
@@ -302,8 +302,10 @@ class HailContext private(val sc: SparkContext,
     nPartitions: Option[Int] = None,
     rg: Option[ReferenceGenome] = Some(ReferenceGenome.defaultReference),
     contigRecoding: Option[Map[String, String]] = None,
-    tolerance: Double = 0.2): MatrixTable = {
-    importBgens(List(file), sampleFile, includeGT, includeGP, includeDosage, nPartitions, rg, contigRecoding, tolerance)
+    tolerance: Double = 0.2,
+    skipInvalidLoci: Boolean = false): MatrixTable = {
+    importBgens(List(file), sampleFile, includeGT, includeGP, includeDosage, nPartitions, rg,
+      contigRecoding, tolerance, skipInvalidLoci)
   }
 
   def importBgens(files: Seq[String],
@@ -314,7 +316,8 @@ class HailContext private(val sc: SparkContext,
     nPartitions: Option[Int] = None,
     rg: Option[ReferenceGenome] = Some(ReferenceGenome.defaultReference),
     contigRecoding: Option[Map[String, String]] = None,
-    tolerance: Double = 0.2): MatrixTable = {
+    tolerance: Double = 0.2,
+    skipInvalidLoci: Boolean = false): MatrixTable = {
 
     val inputs = hadoopConf.globAll(files).flatMap { file =>
       if (!file.endsWith(".bgen"))
@@ -334,7 +337,7 @@ class HailContext private(val sc: SparkContext,
     rg.foreach(ref => contigRecoding.foreach(ref.validateContigRemap))
 
     LoadBgen.load(this, inputs, sampleFile, includeGT: Boolean, includeGP: Boolean, includeDosage: Boolean,
-      nPartitions, rg, contigRecoding.getOrElse(Map.empty[String, String]), tolerance)
+      nPartitions, rg, contigRecoding.getOrElse(Map.empty[String, String]), tolerance, skipInvalidLoci)
   }
 
   def importGen(file: String,
@@ -343,8 +346,9 @@ class HailContext private(val sc: SparkContext,
     nPartitions: Option[Int] = None,
     tolerance: Double = 0.2,
     rg: Option[ReferenceGenome] = Some(ReferenceGenome.defaultReference),
-    contigRecoding: Option[Map[String, String]] = None): MatrixTable = {
-    importGens(List(file), sampleFile, chromosome, nPartitions, tolerance, rg, contigRecoding)
+    contigRecoding: Option[Map[String, String]] = None,
+    skipInvalidLoci: Boolean = false): MatrixTable = {
+    importGens(List(file), sampleFile, chromosome, nPartitions, tolerance, rg, contigRecoding, skipInvalidLoci)
   }
 
   def importGens(files: Seq[String],
@@ -353,7 +357,8 @@ class HailContext private(val sc: SparkContext,
     nPartitions: Option[Int] = None,
     tolerance: Double = 0.2,
     rg: Option[ReferenceGenome] = Some(ReferenceGenome.defaultReference),
-    contigRecoding: Option[Map[String, String]] = None): MatrixTable = {
+    contigRecoding: Option[Map[String, String]] = None,
+    skipInvalidLoci: Boolean = false): MatrixTable = {
     val inputs = hadoopConf.globAll(files)
 
     inputs.foreach { input =>
@@ -371,7 +376,7 @@ class HailContext private(val sc: SparkContext,
 
     //FIXME: can't specify multiple chromosomes
     val results = inputs.map(f => LoadGen(f, sampleFile, sc, rg, nPartitions,
-      tolerance, chromosome, contigRecoding.getOrElse(Map.empty[String, String])))
+      tolerance, chromosome, contigRecoding.getOrElse(Map.empty[String, String]), skipInvalidLoci))
 
     val unequalSamples = results.filter(_.nSamples != nSamples).map(x => (x.file, x.nSamples))
     if (unequalSamples.length > 0)
@@ -476,14 +481,15 @@ class HailContext private(val sc: SparkContext,
     quantPheno: Boolean = false,
     a2Reference: Boolean = true,
     rg: Option[ReferenceGenome] = Some(ReferenceGenome.defaultReference),
-    contigRecoding: Option[Map[String, String]] = None): MatrixTable = {
+    contigRecoding: Option[Map[String, String]] = None,
+    skipInvalidLoci: Boolean = false): MatrixTable = {
 
     rg.foreach(ref => contigRecoding.foreach(ref.validateContigRemap))
 
     val ffConfig = FamFileConfig(quantPheno, delimiter, missing)
 
     LoadPlink(this, bed, bim, fam,
-      ffConfig, nPartitions, a2Reference, rg, contigRecoding.getOrElse(Map.empty[String, String]))
+      ffConfig, nPartitions, a2Reference, rg, contigRecoding.getOrElse(Map.empty[String, String]), skipInvalidLoci)
   }
 
   def importPlinkBFile(bfileRoot: String,
@@ -493,9 +499,10 @@ class HailContext private(val sc: SparkContext,
     quantPheno: Boolean = false,
     a2Reference: Boolean = true,
     rg: Option[ReferenceGenome] = Some(ReferenceGenome.defaultReference),
-    contigRecoding: Option[Map[String, String]] = None): MatrixTable = {
+    contigRecoding: Option[Map[String, String]] = None,
+    skipInvalidLoci: Boolean = false): MatrixTable = {
     importPlink(bfileRoot + ".bed", bfileRoot + ".bim", bfileRoot + ".fam",
-      nPartitions, delimiter, missing, quantPheno, a2Reference, rg, contigRecoding)
+      nPartitions, delimiter, missing, quantPheno, a2Reference, rg, contigRecoding, skipInvalidLoci)
   }
 
   def read(file: String, dropCols: Boolean = false, dropRows: Boolean = false): MatrixTable = {
@@ -584,9 +591,10 @@ class HailContext private(val sc: SparkContext,
     callFields: Set[String] = Set.empty[String],
     rg: Option[ReferenceGenome] = Some(ReferenceGenome.defaultReference),
     contigRecoding: Option[Map[String, String]] = None,
-    arrayElementsRequired: Boolean = true): MatrixTable = {
+    arrayElementsRequired: Boolean = true,
+    skipInvalidLoci: Boolean = false): MatrixTable = {
     importVCFs(List(file), force, forceBGZ, headerFile, nPartitions, dropSamples, callFields, rg, contigRecoding,
-      arrayElementsRequired)
+      arrayElementsRequired, skipInvalidLoci)
   }
 
   def importVCFs(files: Seq[String], force: Boolean = false,
@@ -597,7 +605,8 @@ class HailContext private(val sc: SparkContext,
     callFields: Set[String] = Set.empty[String],
     rg: Option[ReferenceGenome] = Some(ReferenceGenome.defaultReference),
     contigRecoding: Option[Map[String, String]] = None,
-    arrayElementsRequired: Boolean = true): MatrixTable = {
+    arrayElementsRequired: Boolean = true,
+    skipInvalidLoci: Boolean = false): MatrixTable = {
 
     rg.foreach(ref => contigRecoding.foreach(ref.validateContigRemap))
 
@@ -606,7 +615,7 @@ class HailContext private(val sc: SparkContext,
     forceBGZip(forceBGZ) {
       val reader = new HtsjdkRecordReader(callFields)
       LoadVCF(this, reader, headerFile, inputs, nPartitions, dropSamples, rg,
-        contigRecoding.getOrElse(Map.empty[String, String]), arrayElementsRequired)
+        contigRecoding.getOrElse(Map.empty[String, String]), arrayElementsRequired, skipInvalidLoci)
     }
   }
