+++ src/main/scala/is/hail/io/bgen/LoadBgen.scala
@@ -31,7 +31,8 @@ object LoadBgen {
     nPartitions: Option[Int] = None,
     rg: Option[ReferenceGenome] = Some(ReferenceGenome.defaultReference),
     contigRecoding: Map[String, String] = Map.empty[String, String],
-    tolerance: Double): MatrixTable = {
+    tolerance: Double,
+    skipInvalidLoci: Boolean = false): MatrixTable = {
 
     require(files.nonEmpty)
     val hadoop = hc.hadoopConf
@@ -111,26 +112,30 @@ object LoadBgen {
       val rvb = new RegionValueBuilder(region)
       val rv = RegionValue(region)
 
-      it.map { case (_, record) =>
+      it.flatMap { case (_, record) =>
         val (contig, pos, alleles) = record.getKey
         val contigRecoded = contigRecoding.getOrElse(contig, contig)
 
-        rvb.start(kType)
-        rvb.startStruct()
-        rvb.addAnnotation(kType.types(0), Locus.annotation(contigRecoded, pos, rg))
+        if (skipInvalidLoci && !rg.forall(_.isValidLocus(contigRecoded, pos)))
+          None
+        else {
+          rvb.start(kType)
+          rvb.startStruct()
+          rvb.addAnnotation(kType.types(0), Locus.annotation(contigRecoded, pos, rg))
 
-        val nAlleles = alleles.length
-        rvb.startArray(nAlleles)
-        var i = 0
-        while (i < nAlleles) {
-          rvb.addString(alleles(i))
-          i += 1
+          val nAlleles = alleles.length
+          rvb.startArray(nAlleles)
+          var i = 0
+          while (i < nAlleles) {
+            rvb.addString(alleles(i))
+            i += 1
+          }
+          rvb.endArray()
+          rvb.endStruct()
+
+          rv.setOffset(rvb.end())
+          Some(rv)
         }
-        rvb.endArray()
-        rvb.endStruct()
-
-        rv.setOffset(rvb.end())
-        rv
       }
     }))
 
@@ -141,42 +146,46 @@ object LoadBgen {
       val rvb = new RegionValueBuilder(region)
       val rv = RegionValue(region)
 
-      it.map { case (_, record) =>
+      it.flatMap { case (_, record) =>
         val (contig, pos, alleles) = record.getKey
         val va = record.getAnnotation.asInstanceOf[Row]
 
         val contigRecoded = contigRecoding.getOrElse(contig, contig)
 
-        rvb.start(rowType)
-        rvb.startStruct()
-        rvb.addAnnotation(kType.types(0), Locus.annotation(contigRecoded, pos, rg))
-
-        val nAlleles = alleles.length
-        rvb.startArray(nAlleles)
-        var i = 0
-        while (i < nAlleles) {
-          rvb.addString(alleles(i))
-          i += 1
-        }
-        rvb.endArray()
-        rvb.addAnnotation(rowType.types(2), va.get(0))
-        rvb.addAnnotation(rowType.types(3), va.get(1))
-        if (loadEntries)
-          record.getValue(rvb) // gs
+        if (skipInvalidLoci && !rg.forall(_.isValidLocus(contigRecoded, pos)))
+          None
         else {
-          rvb.startArray(nSamples)
-          var j = 0
-          while (j < nSamples) {
-            rvb.startStruct()
-            rvb.endStruct()
-            j += 1
+          rvb.start(rowType)
+          rvb.startStruct()
+          rvb.addAnnotation(kType.types(0), Locus.annotation(contigRecoded, pos, rg))
+
+          val nAlleles = alleles.length
+          rvb.startArray(nAlleles)
+          var i = 0
+          while (i < nAlleles) {
+            rvb.addString(alleles(i))
+            i += 1
           }
           rvb.endArray()
-        }
-        rvb.endStruct()
+          rvb.addAnnotation(rowType.types(2), va.get(0))
+          rvb.addAnnotation(rowType.types(3), va.get(1))
+          if (loadEntries)
+            record.getValue(rvb) // gs
+          else {
+            rvb.startArray(nSamples)
+            var j = 0
+            while (j < nSamples) {
+              rvb.startStruct()
+              rvb.endStruct()
+              j += 1
+            }
+            rvb.endArray()
+          }
+          rvb.endStruct()
 
-        rv.setOffset(rvb.end())
-        rv
+          rv.setOffset(rvb.end())
+          Some(rv)
+        }
       }
     }))
