+++ src/test/resources/skip_invalid_loci.bim
@@ -1,0 +1,5 @@
+0	0:1:A:C	0.0	1	C	A
+1	1:0:A:C	0.0	0	C	A
+1	1:3:A:C	0.0	3	C	A
+1	1:4:A:C	0.0	4	C	A
+1	1:5:A:C	0.0	5	C	A
