+++ src/test/resources/skip_invalid_loci.fam
@@ -1,0 +1,1 @@
+0	0	0	0	0	NA
