+++ src/test/resources/skip_invalid_loci.gen
@@ -1,0 +1,5 @@
+0 0:1:A:C . 1 A C 0 0 0
+1 1:0:A:C . 0 A C 0 0 0
+1 1:3:A:C . 3 A C 0 0 0
+1 1:4:A:C . 4 A C 0 0 0
+1 1:5:A:C . 5 A C 0 0 0
