+++ src/test/resources/skip_invalid_loci.sample
@@ -1,0 +1,3 @@
+ID_1 ID_2 missing
+0 0 0
+0 0 0.0
